Convert location code to db data; correct location hashes
