+++ pkg/models/feedback_source.go
@@ -1,0 +1,10 @@
+package models
+
+import "github.com/jinzhu/gorm"
+
+type FeedbackSource struct {
+	gorm.Model
+	SourceID   string
+	SourceType string
+}
+
