+++ pkg/models/feedback_type.go
@@ -1,0 +1,9 @@
+package models
+
+import "github.com/jinzhu/gorm"
+
+type FeedbackType struct {
+	gorm.Model
+	Description string `gorm:"not null"`
+}
+
