+++ pkg/models/migrate.go
@@ -2,108 +2,7 @@ package models
 
 import (
 	"github.com/jinzhu/gorm"
-	"time"
 )
-
-type Direction struct {
-	gorm.Model
-	Feedback []Feedback
-	Lines    []Line `gorm:"many2many:line_directions"`
-	Name     string `gorm:"not null"`
-}
-
-type Line struct {
-	gorm.Model
-	Feedback   []Feedback  `json:",omitempty"`
-	Directions []Direction `gorm:"many2many:line_directions"`
-	Stations   []Station   `gorm:"many2man:station_lines"`
-	Name       string      `gorm:"not null"`
-}
-
-type Station struct {
-	gorm.Model
-	Feedback []Feedback
-	Detail   StationDetail
-	Lines    []Line `gorm:"many2many:station_lines;not null"`
-	Name     string `gorm:"unique;not null"`
-}
-
-type StationDetail struct {
-	gorm.Model
-	StationID   uint   `gorm:"not null"`
-	Description string `gorm:"not null"`
-	Location    string `gorm:"unique;not null"`
-}
-
-type FeedbackType struct {
-	gorm.Model
-	Description string `gorm:"not null"`
-}
-
-type FeedbackSource struct {
-	gorm.Model
-	SourceID   string
-	SourceType string
-}
-
-type Feedback struct {
-	gorm.Model
-	StationID   int
-	LineID      int
-	DirectionID int
-	SourceID    int
-	Source      FeedbackSource
-	TypeID      int
-	Type        FeedbackType
-	Description string `gorm:"not null"`
-	ThumbsUp    int
-	ThumbsDown  int
-	ExpiresAt   time.Time
-}
-
-// Can't use gorm.Model since we want TrainID as the primary key
-type Train struct {
-	TrainID   uint `gorm:"primary_key:true"`
-	CreatedAt time.Time
-	UpdatedAt time.Time
-	DeletedAt *time.Time
-}
-
-type ScheduleEventSource struct {
-	gorm.Model
-	Name        string `gorm:"not null"`
-	Description string `gorm:"not null"`
-}
-
-type ScheduleEvent struct {
-	gorm.Model
-	EventTypeID   int
-	EventType     ScheduleEventSource
-	DestinationID int
-	Destination   Station
-	NextArrival   time.Time
-	NextStationID int
-	NextStation   Station
-}
-
-type RealTimeEventDetail struct {
-	gorm.Model
-	ScheduleEventID int
-	ScheduleEvent   ScheduleEvent
-	EventTime       time.Time
-	TrainID         int
-	Train           Train
-	WaitingSeconds  int    `gorm:"not null"`
-	WaitingTime     string `gorm:"not null"`
-}
-
-type StaticEventDetail struct {
-	gorm.Model
-	ScheduleEventID    int
-	ScheduleEvent      ScheduleEvent
-	ScheduledTime      time.Time
-	StaticScheduleType string `gorm:"not null"`
-}
 
 func DBMigrate(db *gorm.DB) *gorm.DB {
 	db.LogMode(false)
