+++ pkg/models/schedule_event.go
@@ -1,0 +1,18 @@
+package models
+
+import (
+	"github.com/jinzhu/gorm"
+	"time"
+)
+
+type ScheduleEvent struct {
+	gorm.Model
+	EventTypeID   int
+	EventType     ScheduleEventSource
+	DestinationID int
+	Destination   Station
+	NextArrival   time.Time
+	NextStationID int
+	NextStation   Station
+}
+
