+++ pkg/models/schedule_event_source.go
@@ -1,0 +1,10 @@
+package models
+
+import "github.com/jinzhu/gorm"
+
+type ScheduleEventSource struct {
+	gorm.Model
+	Name        string `gorm:"not null"`
+	Description string `gorm:"not null"`
+}
+
