+++ pkg/models/static_event_detail.go
@@ -1,0 +1,15 @@
+package models
+
+import (
+	"github.com/jinzhu/gorm"
+	"time"
+)
+
+type StaticEventDetail struct {
+	gorm.Model
+	ScheduleEventID    int
+	ScheduleEvent      ScheduleEvent
+	ScheduledTime      time.Time
+	StaticScheduleType string `gorm:"not null"`
+}
+
