+++ pkg/seed/seed.go
@@ -57,10 +57,10 @@ func Seed(db *gorm.DB) {
 	//Multi-line
 	_ = UpsertStation(db, "Lindebergh Center", "dn5bnu0epkt1", "Lindbergh Center Station", []models.Line{gold, red})
 	_ = UpsertStation(db, "Arts Center", "dn5bnu0epkt1", "Arts Center Station", []models.Line{gold, red})
-	_ = UpsertStation(db, "Midtown", "dn5bnu0epkt1", "Midtown Station", []models.Line{gold, red})
-	_ = UpsertStation(db, "North Avenue", "dn5bnu0epkt1", "North Avenue Station", []models.Line{gold, red})
-	_ = UpsertStation(db, "Civic Center", "dn5bnu0epkt1", "Civic Center Station", []models.Line{gold, red})
-	_ = UpsertStation(db, "Peachtree Center", "dn5bnu0epkt1", "Peachtree Center Station", []models.Line{gold, red})
+	_ = UpsertStation(db, "Midtown", "dn5bptxy8r41", "Midtown Station", []models.Line{gold, red})
+	_ = UpsertStation(db, "North Avenue", "dn5bpsp70th7", "North Avenue Station", []models.Line{gold, red})
+	_ = UpsertStation(db, "Civic Center", "dn5bpep496h7", "Civic Center Station", []models.Line{gold, red})
+	_ = UpsertStation(db, "Peachtree Center", "dn5bp9qxs9nh", "Peachtree Center Station", []models.Line{gold, red})
 	_ = UpsertStation(db, "Five Points", "dn5bp8ezwy4k", "Five Points Stations", []models.Line{gold, red, blue, green})
 	_ = UpsertStation(db, "Garnett", "djgzzxbb2xkd", "Garnett Station", []models.Line{gold, red})
 	_ = UpsertStation(db, "West End", "djgzzjeb581t", "West End Station", []models.Line{gold, red})
