--- pkg/transformers/transformers_data/station_locations.go
@@ -1,161 +1,0 @@
-package transformers_data
-
-type GeoLocation struct {
-	Name     string
-	Location string
-}
-
-var GeoLocations = []GeoLocation{
-	{
-		Name:     "North Springs Station",
-		Location: "dnh107scwx23",
-	},
-	{
-		Name:     "Sandy Springs Station",
-		Location: "dnh10939s87f",
-	},
-	{
-		Name:     "Dunwoody Station",
-		Location: "dnh0bxnr3hcj",
-	},
-	{
-		Name:     "Medical Center Station",
-		Location: "dnh0bt32f0zr",
-	},
-	{
-		Name:     "Buckhead Station",
-		Location: "dnh084sbc4fj",
-	},
-	{
-		Name:     "Doraville Station",
-		Location: "dnh0f5v6mxzj",
-	},
-	{
-		Name:     "Chamblee Station",
-		Location: "dnh0c94gqrm2",
-	},
-	{
-		Name:     "Brookhaven Station",
-		Location: "dnh08u1fpng1",
-	},
-	{
-		Name:     "Lenox Station",
-		Location: "dnh0837g7frm",
-	},
-	{
-		Name:     "Lindbergh Center Station",
-		Location: "dnh02jk56z8d",
-	},
-	{
-		Name:     "Arts Center Station",
-		Location: "dn5bpxpn2jvw",
-	},
-	{
-		Name:     "Midtown Station",
-		Location: "dn5bptxy8r41",
-	},
-	{
-		Name:     "North Avenue Station",
-		Location: "dn5bpsp70th7",
-	},
-	{
-		Name:     "Civic Center Station",
-		Location: "dn5bpep496h7",
-	},
-	{
-		Name:     "Peachtree Center Station",
-		Location: "dn5bp9qxs9nh",
-	},
-	{
-		Name:     "Five Points Station",
-		Location: "dn5bp8ezwy4k",
-	},
-	{
-		Name:     "Garnett Station",
-		Location: "djgzzxbb2xkd",
-	},
-	{
-		Name:     "West End Station",
-		Location: "djgzzjeb581t",
-	},
-	{
-		Name:     "Oakland City Station",
-		Location: "djgzyf5dfsmn",
-	},
-	{
-		Name:     "Lakewood Station",
-		Location: "djgzwz0c6suf",
-	},
-	{
-		Name:     "East Point Station",
-		Location: "djgzwdb63g2k",
-	},
-	{
-		Name:     "College Park Station",
-		Location: "djgzqq4k3j73",
-	},
-	{
-		Name:     "Airport Station",
-		Location: "djgzqkhjse84",
-	},
-	{
-		Name:     "Hamilton E. Holmes Station",
-		Location: "dn5bjbfgcmr3",
-	},
-	{
-		Name:     "Bankhead Station",
-		Location: "dn5bnu0epkt1",
-	},
-	{
-		Name:     "West Lake Station",
-		Location: "dn5bn2sgc1bc",
-	},
-	{
-		Name:     "Ashby Station",
-		Location: "dn5bp11qp0s9",
-	},
-	{
-		Name:     "Vine City Station",
-		Location: "dn5bp34zmh5s",
-	},
-	{
-		Name:     "Omni Station",
-		Location: "dn5bp90pezjh",
-	},
-	{
-		Name:     "Georgia State Station",
-		Location: "dn5bp8pgdtcf",
-	},
-	{
-		Name:     "King Memorial Station",
-		Location: "dn5bpbp8fe6s",
-	},
-	{
-		Name:     "Inman Park Station",
-		Location: "dnh0092w0nxh",
-	},
-	{
-		Name:     "Edgewood-Candler Park Station",
-		Location: "dnh00f1wzrc6",
-	},
-	{
-		Name:     "East Lake Station",
-		Location: "dnh016v1ynv9",
-	},
-	{
-		Name:     "Decatur Station",
-		Location: "dnh01u9cru4h",
-	},
-	{
-		Name:     "Avondale Station",
-		Location: "dnh04heebfzp",
-	},
-	{
-		Name:     "Kensington Station",
-		Location: "dnh04u1s7ycg",
-	},
-	{
-		Name:     "Indian Creek Station",
-		Location: "dnh0579u6fcg",
-	},
-}
